Fix a compilation warning (simulator build on 64-bit host).
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24251 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/usbstack/usb_hid.c
@@ -551,7 +551,7 @@ static size_t descriptor_report_get(unsigned char *dest)
     PACK_VAL(report, END_COLLECTION);
 #endif /* HAVE_USB_HID_MOUSE */
 
-    return (size_t)((uint32_t)report - (uint32_t)dest);
+    return (size_t)(report - dest);
 }
 
 static void descriptor_hid_get(unsigned char **dest)
